Inherit dtype when possible to ease potential dtype change in the future (#46)

+++ opennmt/decoders/decoder.py
@@ -16,7 +16,7 @@ def logits_to_cum_log_probs(logits, sequence_length):
   Returns:
     The cumulated log probability of each sequence.
   """
-  mask = tf.sequence_mask(sequence_length, dtype=tf.float32)
+  mask = tf.sequence_mask(sequence_length, dtype=logits.dtype)
   mask = tf.expand_dims(mask, -1)
 
   log_probs = tf.nn.log_softmax(logits)
